refactor: publish workspace snapshots through Effect
