+++ package.json
@@ -15,7 +15,7 @@
     "effect:inventory": "node scripts/effect-backend-inventory.mjs",
     "effect:inventory:write": "node scripts/effect-backend-inventory.mjs --write",
     "effect:architecture": "node scripts/verify-effect-architecture.mjs",
-    "test:effect-backend": "tsx --test proof/plan-program.test.ts proof/terminal-gh-probe-program.test.mjs proof/terminal-live-client-program.test.mjs proof/terminal-upgrade-probe-program.test.mjs src/conversation-starters.test.ts src/cycle-costs.test.ts src/decision-response.test.ts src/desk-write.test.ts src/dead-session.test.ts src/dead-session-scan.test.ts src/gateway-retry-fetch.test.ts src/mcp-probe.test.ts src/notify.test.ts src/pantry-client.test.ts src/reusable-tool-preferences.test.ts src/session-pinning.test.ts src/voice-check-prompt.test.ts src/web-search.test.ts src/workspace-mcp.test.ts && npm run test:lifecycle-agents",
+    "test:effect-backend": "tsx --test proof/plan-program.test.ts proof/terminal-gh-probe-program.test.mjs proof/terminal-live-client-program.test.mjs proof/terminal-upgrade-probe-program.test.mjs src/conversation-starters.test.ts src/cycle-costs.test.ts src/decision-response.test.ts src/desk-write.test.ts src/dead-session.test.ts src/dead-session-scan.test.ts src/gateway-retry-fetch.test.ts src/mcp-probe.test.ts src/notify.test.ts src/pantry-client.test.ts src/reusable-tool-preferences.test.ts src/session-pinning.test.ts src/voice-check-prompt.test.ts src/workspace-snapshot-ordering.test.ts src/web-search.test.ts src/workspace-mcp.test.ts && npm run test:lifecycle-agents",
     "build:effect-workers": "npm run build:assets && wrangler deploy --dry-run --env= --outdir .wrangler/effect-main && wrangler deploy --dry-run --env= --config agents/wrangler.jsonc --outdir .wrangler/effect-factory && wrangler deploy --dry-run --env= --config agents/wrangler.hook.jsonc --outdir .wrangler/effect-factory-hook",
     "verify:effect-backend": "npm run effect:inventory && npm run effect:architecture && npm run test:effect-backend && npm run build:effect-workers && npm run verify:public",
     "check": "npm run build:assets && npm run typecheck && npm run test:unit && npm run test:lifecycle-agents",
